requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.2.2
django-filter==23.2
drf-spectacular>=0.25,<1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
requests==2.31.0
Pillow>=9.0
boto3>=1.26
django-storages>=1.12,<1.15
django-redis>=5.2
moto[s3]>=4.0
\n
